requirements.txt
asgiref==3.8.1
cmake==3.30.3
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
markdown2==2.4.13
pillow==10.4.0
platformdirs==4.2.2
PyJWT==2.9.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
virtualenv==20.26.3
